tool_serch_tool/pyproject.toml
[project]
name = "tool-serch-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.2",
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-core>=1.2.5",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "mcp>=1.25.0",
    "numpy>=2.2.6",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "spotipy>=2.25.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "unicorn>=2.1.4",
    "uvicorn>=0.40.0",
]
